Rewrite, cleanup, renaming
