+++ checks/apps/neko/src/32768/tgv.case
@@ -1,0 +1,50 @@
+{
+  "version": 1.0,
+  "case":
+  {
+  "mesh_file": "32768.nmsh",
+  "output_boundary": true,
+  "output_checkpoints": false,
+  "output_at_end": true,
+  "load_balance": false,
+  "job_timelimit": "00:10:00",
+  "end_time": 2.0,
+  "timestep": 5d-4,
+  "numerics": {
+    "time_order": 3,
+    "polynomial_order": 7,
+    "dealias": true
+  },
+  "fluid": {
+    "scheme": "pnpn",
+    "Re": 1600,
+    "initial_condition": {
+      "type": "user"
+    },
+    "velocity_solver": {
+      "type": "cg",
+      "preconditioner": "jacobi",
+      "projection_space_size": 0,
+      "absolute_tolerance": 1e-7,
+      "max_iterations": 800
+    },
+    "pressure_solver": {
+      "type": "gmres",
+      "preconditioner": "hsmg",
+      "projection_space_size": 20,
+      "absolute_tolerance": 1e-7,
+      "max_iterations": 800
+    },
+    "output_control": "simulationtime",
+    "output_value": 1 
+  },
+  "simulation_components": 
+  [
+    {
+      "type": "vorticity",
+      "compute_control": "tsteps",
+      "compute_value": 50
+    }
+  ]
+  }
+}
