--- checks/apps/neko/src/rocm_select_gpu_device
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-export ROCR_VISIBLE_DEVICES=${SLURM_LOCALID}
-exec $*
